requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pymongo
dnspython
python-dotenv
langchain
